Baseline for Shoal0.1 pypi packages. closes #9
setup files will check if user is root or not, and install files into
the appropriate places.

All setup files work, recommended to use sudo and system wide install to install easily, but can
install into a virtualenv as well, but will need to modify some settings for it to work properly.

fixed some typos.

config files can be located in ~/.shoal/ or /etc/shoal/ depending on
priviledge level on setup.

shoal_server will be located in users home directory if not root when
installing shoal server, otherwise in /var/shoal/.