Shoal-Agent is now more helpful on errors.
Added two conf settings, external_ip and tx_bytes_path.

External_ip is used if no public IP can be found, and must be set if the
squid has no public IP or the message will not send.

tx_bytes_path must be set to the proper path of the NIC we want to use
for load data.

updated micro version.