Adding the phantom sensors to be used with shoal in the repository.
The decision engine (DE) script uses the HTTP Api of phantom to control the domain. It uses the shoal server stats to gather total load of all connected agents. This controls a multi-VM domain.

The txKBps script examines the output bytes of a single shoal-agent and is meant to be used with phantom's autoscale web interface. This is used by a sensor domain to scale the number of VMs