Fix Issue #33
shoal-client requires compatibility with python 2.4 with no external
libraries. Previously, shoal-client used the simplejson (incoporated into
python in 2.6) library. This commit introduces a first pass for a parser
of shoal-server data. It is not a proper JSON parser; it only searches
for specific keys in the data. The list of keys should probably be moved
into a config file that both the server and client use. This commit also
effects Issue #30 and Commit fefa482e3483a454e8d1b8fee0b380a274ccfb57 as
with open(..) as is not available in 2.4.