Refactor of the config parser code to prevent shadowing errors
The way the config.py files were done before were prone to shadowing
errors; every variable was declared globally to make it acessible to the
shoal branch main code and to set default values. The default was overwritten in
the setup function which, had global declerations for all variable at
the start. Forgetting to add a global decleration caused shadowing; both a
local and a variable were created with the same name. The default value
was never overwritten in setup because a new local variable of the same
name was created. To prevent this, the setup function was removed as its
parameter was not used.

Also minor updates to the default config files (removed some outdated
comments, and changed default AMQP server to be shoal.heprc.uvic.ca)