Fix log strings and IP bug
Fixed logging strings in utilities.py that were causing the thread to
die unexpectedly when the agent fails to advertise some of the squid
properties. (mainly public/private ip)

Fixed some logic in the json unpacking where the server would fail
to set the public ip to the external ip when no public ip was given.