Add timeout to verification #87
I've added a 2 second timeout to the verification request.
This should allow time for the file to be downloaded even
if it is half the world away. This should make for a more
responsive server on a restart adding the squids back to
human UI much faster. This will also make it so the server
never spends more than 2 seconds on a bad(unverifiable)
squid.