Add script to generate agent rpm
I've added a script that has some basic functionality to build
agent rpms from a git checkout. Currently it is slightly
customizable through passing arguments which are outlined
in the comments of the file. In the future this script
should be updated to use flags so that the order of arguments
do not pose a problem. Looking forward this script should be
called by a higher level one that can manage all three shoal
components.