Remove verification from consumer thread
In this commit I've changed setup.py such that none of
the verification happens outside the verification thread.
I've also included the timeout change from the clockdrift
branch. #88

These changes also result in changes to the utilities
file where the verification function actually runs.
the verify_new_squid function has been removed since
it will always be running the verify function instead.

The final change is to the index template file. The
change reflects the changes that have already been
made to the production server such that squids of
the "same domain only" class are displayed as green
even when they are not verified. This has no affect
on the functionality of shoal and is only reflected
in the human readable web interface. In the future
when changes are made to the domain_access and
global_access booleans this change should be
reverted since squids from that class should be
able to be verified by the server.