Build server from spec file
This commit moves a lot of files around so they can
be properly built with a spec file as per redhat standard.
the spec file has been tested with centos6 and runs
smoothly after a yum localinstall with the rpm generated.
Some of the dependencies are in the shoal repo so it is
neccisairy to curl in the repo before trying to install
or you wont be able to find some of the dependencies.

The documentation on github is still fairly accurate
for an install with the rpm generated with this spec file
but should probably be double checked once we have a new
rpm in the repo.