Update default proxy behaviour
This commit chagnes the client to read the enviroment variable and config file
and write them both (if they exist) to the http proxy.
This also means that if it fails to reach the shoal server it doesn't just give up.
Instead, it logs some information about the proxys set and then exits nicely after
setting the defaults.

Tested on sl7.