add timeout to shoal-server request
During testing i noticed that if there was a bad proxy preconfigured
as the HTTP_PROXY it would take a long time to timeout. The timeout should
change that but there is still a problem when making a requst to the
shoal-server with the bad proxy as it will never actually reach it.