update error handling of geolite city db download
If the database file is manually updated in a way where the apache user loses write access
then the auto download/update function throws an error crashing the thread.

This change simply catches and reports the error such that execution can continue.
We may want to upgrade this routine in the future to check if the database truly
needs to be updated before writing over the old database file that may have some
manual insertions.