change basic shoal client functionality and add new force-server flag
This commit is untested but attmepts to change the order of operations in the shoal client
to always try to wait for a local broadcast before contacting the shoal-server. If a squid
is found via broadcast it will forgo contacting the shoal server unless you explicitly use
the --force-server or -o flag.